Do not allow negative content-lengths

+++ mk_http_parser.h
@@ -114,8 +114,8 @@ struct mk_http_parser {
     int chars;
 
     /* it stores the numeric value of Content-Length header */
-    size_t body_received;
-    size_t header_content_length;
+    long body_received;
+    long header_content_length;
 
     /* probable current header, fly parsing */
     int header_key;
